fix: изменить доступ к удалению клиента

+++ src/main/java/ru/modgy/owner/controller/OwnerController.java
@@ -95,6 +95,7 @@ public class OwnerController {
             @PathVariable("id") Long id
     ) {
         log.info("OwnerController: requesterId={} DELETE/deleteOwnerById id= {}", requesterId, id);
+        utilityService.checkBossAccess(requesterId);
         ownerService.deleteOwnerById(requesterId, id);
     }
